% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_requirements.R
\name{profile_required_indicators}
\alias{profile_required_indicators}
\title{Get Required Indicators for a Profile}
\usage{
profile_required_indicators(profile)
}
\arguments{
\item{profile}{A \code{submission_profile} object.}
}
\value{
A character vector of required \code{indicator_id} values.
}
\description{
Returns the list of indicator IDs that are mandatory for the given
submission profile.
}
\examples{
prof <- submission_profile("FDA", "NDA")
profile_required_indicators(prof)

prof_ind <- submission_profile("FDA", "IND")
profile_required_indicators(prof_ind)

}
