% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_configs.R
\name{profile_risk_config}
\alias{profile_risk_config}
\title{Extract Risk Configuration from a Profile}
\usage{
profile_risk_config(profile)
}
\arguments{
\item{profile}{A \code{submission_profile} object.}
}
\value{
A list of class \code{"risk_config"} with \code{rpn_bands},
\code{evidence_severity_to_probability}, \code{evidence_severity_to_impact},
and \code{default_detectability}.
}
\description{
Returns a list with structure compatible with \code{r4subrisk::risk_config_default()}.
Can be passed directly to risk assessment functions.
}
\examples{
prof <- submission_profile("FDA", "NDA")
cfg <- profile_risk_config(prof)
cfg$rpn_bands

}
